src/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
# Editable Git install with no remote (aurora==0.1.0)
-e c:\users\winha\aurora\aurora-platform
aurora_crm==0.1.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
flake8==7.2.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pyotp==2.9.0
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
qrcode==8.2
redis==6.2.0
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
